Added API 21 hookAutoIncrement.
Due to package name changes between KitKat and Lollipop, StatusBarManagerService moved from com.android.server to com.android.server.statusbar.
This is why it could not be found by Lollipop devices.
Also, the method updateNotification is removed on Lollipop.
Solution for this has to be found, i.e. alternative possible method for this hook.